multi-ret: don't use $code->{auto}
OK, lets make the change mentioned in the previous commit. Just to add
more turbulence to test/ ;) The only additional change is s/auto/supp/
for clarity.

So. AUTO/auto no longer have internal users, this feature was added for
multi-ret before the out-of-band 'supp' logic was fully implemented.

Should it die? probably, but perhaps I will find the "natural" use-case
for this feature in the external .dsp code, so lets keep it for now.