parse_fpp: simplify the usage of ret-type == $type[0]
While at it, delete $code->{stmt}, just to make it clear that it is
no longer used.

TODO: if stmt/postp != '', ret-type passed to mk_ffunc() before return
has no effect, but unless it is _(int|float)_ it abuses $func->{type}
checked in fix_types(). Not a real problem but doesn't look good.